+++ lib/pull-request-handler.js
@@ -1,7 +1,7 @@
-const handlerSetup = require('./handler-setup')
-const generateAssignedTo = require('./generate-assigned-to')
+const handlerSetup = require('./utils/handler-setup')
+const generateAssignedTo = require('./utils/generate-assigned-to')
+const { truncate } = require('./utils/helpers')
 const { comment } = require('./templates')
-const { truncate } = require('./helpers')
 
 module.exports = async context => {
   const [
