+++ lib/pull-request-merged-handler.js
@@ -1,8 +1,8 @@
-const generateAssignedTo = require('./generate-assigned-to')
-const checkForDuplicateIssue = require('./check-for-duplicate-issue')
-const handlerSetup = require('./handler-setup')
+const generateAssignedTo = require('./utils/generate-assigned-to')
+const checkForDuplicateIssue = require('./utils/check-for-duplicate-issue')
+const handlerSetup = require('./utils/handler-setup')
+const { truncate } = require('./utils/helpers')
 const { issue } = require('./templates')
-const { truncate } = require('./helpers')
 
 module.exports = async context => {
   const [
