+++ lib/utils/default-config.js
