+++ lib/utils/reopen-closed.js
